// b2c-app/src/app/booking/models/vocuher-info.ts
export class voucherInfo {
  code: string;
  data: any;
  paymentData: any;
  email: string;
  mobileNo: string;
  businessLoggedOnToken: string;
  contactDetails: contactData;
}

export class contactData {
  mobileCode: string;
  mobileCodeKey: string;
  title: string;
  firstName: string;
  lastName: string;
}